LayoutTests/media/video-pause-empty-events-expected.txt

Test that pause() from EMPTY network state triggers load()

EXPECTED (video.networkState == '3') OK
RUN(video.pause())
SCRIPT DONE
EVENT(loadstart)
EVENT(durationchange)
EVENT(loadedmetadata)
EVENT(loadeddata)
EVENT(canplay)
EVENT(canplaythrough)
END OF TEST
